Component name refactoring; avoid reserved names
