+++ client/src/components/History/Layout/LayoutDetails.vue
